Notify orderpool cleanup without IPC (#327)
## 📝 Summary

Closes #318. This PR uses the block building flow to notify the order
pool to clean old transactions instead of using the IPC path of the EL
node.

## 💡 Motivation and Context

<!--- (Optional) Why is this change required? What problem does it
solve? Remove this section if not applicable. -->

---

## ✅ I have completed the following steps:

* [ ] Run `make lint`
* [ ] Run `make test`
* [ ] Added tests (if applicable)

---------
